+++ baybe/utils/__init__.py
@@ -8,6 +8,7 @@ import pickle
 
 import ssl
 import urllib.request
+from abc import ABC
 from functools import lru_cache
 from pathlib import Path
 
@@ -43,6 +44,28 @@ memory_utils = Memory(cachedir / "utils")
 
 # Global Mordred calculator (variable could be replaced with a singleton pattern)
 mordred_calculator = Calculator(descriptors)
+
+
+def isabstract(cls: Any) -> bool:
+    """
+    Determines if a given class is abstract in a more general sense than
+    `inspect.abstract`, which only verifies if a class has abstract methods. The
+    latter can be problematic when the class has no abstract methods but is
+    nevertheless not directly usable, for example, because it has uninitialized
+    members, which are only covered in its non-"abstract" subclasses. By contrast,
+    this method simply checks if the class derives from `abc.ABC`.
+
+    Parameters
+    ----------
+    cls : Any
+        The class to be inspected.
+
+    Returns
+    -------
+    bool
+        True if the class is "abstract" (see definition above), False else.
+    """
+    return ABC in cls.__bases__
 
 
 def is_valid_smiles(smiles: str) -> bool: